SyntheticEddyMethod.jl

Package for synthetic eddy generation

Package Features

Create coeherent eddies in 3D domain

Function Documentation

Missing docstring.

Missing docstring for SyntheticEddyMethod. Check Documenter's build log for details.

SyntheticEddyMethod.SEM_EDDY — Type

Structure with the property of a single eddy

SyntheticEddyMethod.VirtualBox — Type

Virtual Volume box where the eddies are created

SyntheticEddyMethod.convect_eddy — Function

Compute the new position of all the Eddies. We consider only the convective velocity along x axis. If outside the Virtual Box, a new eddy is randomly generated inside the Virtual Box

Missing docstring.

Missing docstring for create_vector_points. Check Documenter's build log for details.

SyntheticEddyMethod.compute_U_k — Function

Compute the acutual velocity and the turbulent kinetic energy. The convective velocity is just in the x direction

SyntheticEddyMethod.initialize_eddies — Function

Initialize Eddy position and intensity

SyntheticEddyMethod.compute_uᵢₚ — Function

The velocity in the 3 directions is computed in each point provided in x