Exported Types and Functions

SyntheticEddyMethod.SEM_EDDY — Type

Structure with the property of a single eddy

SyntheticEddyMethod.VirtualBox — Type

Virtual Volume box where the eddies are created

SyntheticEddyMethod.convect_eddy — Function

Compute the new position of all the Eddies. We consider only the convective velocity along x axis. If outside the Virtual Box, a new eddy is randomly generated inside the Virtual Box

SyntheticEddyMethod.create_vector_points — Function

Create a vector of points

SyntheticEddyMethod.compute_Ek — Function

Compute turbulent kinetic energy

SyntheticEddyMethod.initialize_eddies — Function

Initialize Eddy position and intensity

SyntheticEddyMethod.compute_uᵢₚ — Function

The velocity in the 3 directions is computed in each point provided in x

SyntheticEddyMethod.cholesky_decomposition — Function

Cholesky Decomposition of Reynolds Stress Tensor