Index

Shape Functions

SyntheticEddyMethod.tent_fun — Function

tent_fun(x)

Tent-like shape function. The domain of the function is [-1,1]x[-1,1]x[-1,1].
It satisfy the normalization condition:
$\int_{-1}^{1} fσ^2(x) dx = 1$

SyntheticEddyMethod.step_fun — Function

step_fun(x)

Step function. The domain of the function is [-1,1]x[-1,1]x[-1,1].
It satisfy the normalization condition:
$\int_{-1}^{1} fσ^2(x) dx = 1$

SyntheticEddyMethod.trunc_gauss_fun — Function

trunc_gauss_fun(x)

Truncated Gaussian function. The domain of the function is [-1,1]x[-1,1]x[-1,1].
It satisfy the normalization condition:
$\int_{-1}^{1} fσ^2(x) dx = 1$

Decomposition

SyntheticEddyMethod.cholesky_decomposition — Function

cholesky_decomposition(R::Matrix{Float64})

Cholesky Decomposition of the Reynolds Stress Tensor.

Reynolds Stress

SyntheticEddyMethod.Reynolds_stress_tensor — Type

Reynolds_stress_tensor

Struct hosting the Reynolds stress tensor informations

SyntheticEddyMethod.Reynolds_stress_points — Function

Reynolds_stress_points(vec_points::Vector{Vector{Float64}}, A::Reynolds_stress_interpolator)

Create vecA of type Vector{Matrix}, where length(vecA) == length(vec_points) Each element is the cholesky decomposition of the Reynolds stress at the specific point

Reynolds_stress_points(vec_points::Vector{Vector{Float64}}, A::Matrix{Float64})

Create vecA of type Vector{Matrix}, where length(vecA) == length(vec_points) Each element is the cholesky decomposition of the Reynolds stress at the specific point.

SyntheticEddyMethod.Reynolds_stress_point — Function

Reynolds_stress_point(point::Vector{Float64}, Re::Reynolds_stress_interpolator)

It gives the Reynolds stress for a specific point. It performs an interpolation using the information in Re.

SyntheticEddyMethod.get_reynolds_stress_from_file — Function

get_reynolds_stress_from_file(Re_file_info::String)

Function called by the user, where the Refileinfo is path of the .xlsx file with the data of the Reynolds Stress.
File column example:
Z | Y | UU | VV | WW | UV | UW | VW

Eddies

SyntheticEddyMethod.AbstractEddy — Type

AbstractEddy

Abstract type defining an Eddy object.

SyntheticEddyMethod.SemEddy — Type

SemEddy

It identify the properties of each eddy.

eddy_num::Int64 : Eddy identification number

σ::Vector{Float64} : Eddy length scale

xᵢ::Vector{Float64} : Eddy's position in the computational box [x,y,z]

ϵᵢ::Vector{Float64} : Eddy's intensity (+1 or -1) in [x,y,z]

SyntheticEddyMethod.VirtualBox — Type

VirtualBox

Virtual Volume box where the eddies are created.

σ::Vector{Float64} : 3 elements vector, to have different σ in different directions

N::Int64 : number of eddies

shape_fun::Function : shape function

V_b::Float64 : volume of the virtual box

X_start::Float64

X_end::Float64

Y_start::Float64

Y_end::Float64

Z_start::Float64

Z_end::Float64

SyntheticEddyMethod.convect_eddy — Function

convect_eddy(dt::Float64, Eddy::SemEddy, U₀::Float64, σ::Vector{Float64}, Vbinfo::VirtualBox)

The eddies are convected, shifted in the x direction of a distance dt * U₀.
If one eddy exits the Virtual Box, it is re-generated randomly inside the Virtual Box.

SyntheticEddyMethod.create_vector_points — Function

create_vector_points(x, y, z)

Create a vector of points. Useful for testing purposes.

Examples

julia> create_vector_points([1.0], [2.0, 3.0], [1.5, 3.5, 4.2])
6-element Vector{Vector{Float64}}:
 [1.0, 2.0, 1.5]
 [1.0, 2.0, 3.5]
 [1.0, 2.0, 4.2]
 [1.0, 3.0, 1.5]
 [1.0, 3.0, 3.5]
 [1.0, 3.0, 4.2]

SyntheticEddyMethod.initialize_eddies — Function

initialize_eddies(Vbinfo::VirtualBox)

Initialize Eddy position and intensity

SyntheticEddyMethod.new_rand_position — Function

new_rand_position(Vbinfo::VirtualBox)

It computes a random position inside the Virtual Box volume.

SyntheticEddyMethod.uᵢ — Function

uᵢ(vec_points::Vector{Vector{Float64}}, 
ϵᵢ::Float64, xᵢ::Vector{Float64}, σ::Vector{Float64}, 
f::Function)

For each i point in vec_points it computes:
$f\left(\frac{x_1-x_1^k}{\sigma_1}\right)f\left(\frac{x_2-x_2^k}{\sigma_2}\right) f\left(\frac{x_3-x_3^k}{\sigma_3}\right)$

SyntheticEddyMethod.compute_fluct — Function

compute_fluct(vec_points::Vector{Vector{Float64}}, dt::Float64, Eddies::Vector{SemEddy}, U₀::Float64, Vbinfo::VirtualBox, Re::Union{Matrix,Reynolds_stress_interpolator})

Compute the velocity fluctuations accordingly to the Reynolds Stress Re.

SyntheticEddyMethod.compute_uᵢₚ — Function

compute_uᵢₚ(x::Vector{Vector{Float64}}, dt::Float64, Eddies::Vector{SemEddy}, U₀::Float64, Vbinfo::VirtualBox)

The velocity in the 3 directions is computed in each point provided in vec_points.

SyntheticEddyMethod.compute_Ek — Function

compute_Ek(U::Vector{Vector{Float64}}, U₀::Float64)

Compute the turbulent kinetic energy. The convective speed U₀ is subtracted from the x component of the speed.

Spectral Utilities

SyntheticEddyMethod.fft_from_signal — Function

fft_from_signal(q::Vector{Float64},dt::Float64)

It performs the Fast Fourier transformation of the signal q.

Examples

# We create a random signal
julia> q = rand(1000)
1000-element Vector{Float64}:
 0.7478100061738703
 0.16303784021483914
 0.3628099523805166
 0.2608705413573018
 0.22498693840252693
 0.9428769056802648
 0.3579787963511998
 0.7224804012744513
 ⋮
 0.9946122085650284
 0.9183278219776294
 0.19908223170935013
 0.7708756376182156
 0.5007439030983675
 0.8464298265686861
 0.8700496116560734
 0.3670620181980453
julia> psd, freq = fft_from_signal(q,0.1)
([0.11064559574015868, 0.005389219813554343, 0.21634910063286397, 0.12896616563010807, 0.1578838809521174, 0.051975560320304294, 0.10663607461615124, 0.1119872950890813, 0.009687265867758152, 0.1469444321528952  …  0.08051345232908684, 0.1481450312902375, 0.018682727272637666, 0.19356783851443643, 0.006863014095143317, 0.009106381063590546, 0.01426246439106842, 0.25854085870704596, 0.08248723638591579, 0.18281921062421522], [0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09, 0.1  …  4.9, 4.91, 4.92, 4.93, 4.94, 4.95, 4.96, 4.97, 4.98, 4.99])
julia> length(psd)
499
julia> length(freq)
499